Move the bounding box down into the shared layer
