+++ src/lib/shared/bounding-box/index.ts
@@ -1,4 +1,4 @@
-export type { BBox, RelativeBBox } from './model/type';
+export type { BBox, RelativeBBox } from './lib/type';
 export { expandBBox } from './lib/expand-bbox';
 export { getPathBBox } from './lib/get-path-bbox';
 export { inscribeInSquare } from './lib/inscribe-in-square';
