+++ src/lib/shared/bounding-box/lib/type.d.ts
